DEV-2507 Add tests for basic 1.2 profile
Remove BasicProfile, the profile version 1.2 will now inherit from 1.1.

Add some fixes in the SPAQRL files.
