+++ app/models/profiles/profile.py
@@ -9,7 +9,6 @@ from pysparql_anything import SparqlAnything
 
 from app.models.profiles.exceptions import (
     GraphNotConformError,
-    GraphParseError,
     XMLNotValidError,
 )
 
@@ -190,7 +189,7 @@ class Profile(ABC):
             GraphNotConformError: If not conform, containing the reason why.
         """
         # An empty graph conforms with the SHACL. Check if the graph has at least
-        # a premis:intellecutalEntity node.
+        # a premis:intellectualEntity node.
         if (
             None,
             URIRef("http://www.w3.org/1999/02/22-rdf-syntax-ns#type"),
@@ -211,114 +210,3 @@ class Profile(ABC):
             raise GraphNotConformError(results_text)
 
         return True
-
-
-class BasicProfile(Profile):
-
-    @abstractmethod
-    def dcterms_xsd() -> Path:
-        pass
-
-    def _validate_descriptive(self) -> list[XMLNotValidError]:
-        """Validate the dcterms file.
-
-        Basic profile has one file with the descriptive metadata.
-
-        Returns:
-            A parse/validation error in a list if applicable.
-        """
-        dcterms_xsd = etree.XMLSchema(etree.parse(self.dcterms_xsd()))
-
-        dcterms_package_path: Path = self.bag_path.joinpath(
-            "data",
-            "metadata",
-            "descriptive",
-            "dc_1.xml",
-        )
-
-        # DCTERMS on package level
-        try:
-            dcterms_package = etree.parse(dcterms_package_path)
-            dcterms_xsd.assertValid(dcterms_package)
-        except (etree.DocumentInvalid, etree.ParseError) as e:
-            return [XMLNotValidError(str(e.error_log))]
-
-        return []
-
-    def parse_graph(self) -> Graph:
-        """Parse the metadata as a graph.
-
-        As the SPARQL service paths are dynamic, we need to instantiate
-        the SPARQL file and write it in the bag folder.
-
-        Transform the metadata to rdf and then load it into a graph.
-
-        Returns:
-            The graph.
-        Raises:
-            GraphParseError: If parsing failed.
-        """
-        # write SPARQL-anything query for the SIP.
-        query_sip_destination = self.bag_path.joinpath(self.query_sip().name)
-        sip_template = self.jinja_env.get_template(str(self.query_sip().name))
-        with open(str(query_sip_destination), "w") as f:
-            f.write(sip_template.render(bag_path=self.bag_path))
-
-        # write SPARQL-anything query for the descriptive metadata on IE level.
-        query_descriptive_ie_destination = self.bag_path.joinpath(
-            self.query_descriptive_ie().name
-        )
-        profile_template = self.jinja_env.get_template(
-            str(self.query_descriptive_ie().name)
-        )
-        with open(str(query_descriptive_ie_destination), "w") as f:
-            f.write(profile_template.render(bag_path=self.bag_path))
-
-        # write SPARQL-anything query for the premis metadata on the IE level.
-        query_premis_ie_destination = self.bag_path.joinpath(
-            self.query_premis_ie().name
-        )
-        profile_template = self.jinja_env.get_template(str(self.query_premis_ie().name))
-        with open(str(query_premis_ie_destination), "w") as f:
-            f.write(profile_template.render(bag_path=self.bag_path))
-
-        # write SPARQL-anything query for the premis metadata on the representation level.
-        query_premis_rep_destination = self.bag_path.joinpath(
-            self.query_premis_representation().name
-        )
-        profile_template = self.jinja_env.get_template(
-            str(self.query_premis_representation().name)
-        )
-        with open(str(query_premis_rep_destination), "w") as f:
-            f.write(profile_template.render(bag_path=self.bag_path))
-
-        # Run SPARQL-anything transformation.
-        try:
-            sip_graph = sa.construct(q=str(query_sip_destination), f="TTL")
-            profile_descriptive_ie = sa.construct(
-                q=str(query_descriptive_ie_destination), f="TTL"
-            )
-            profile_premis_ie = sa.construct(
-                q=str(query_premis_ie_destination), f="TTL"
-            )
-            profile_premis_rep = sa.construct(
-                q=str(query_premis_rep_destination), f="TTL"
-            )
-
-            data_graph = (
-                sip_graph
-                + profile_descriptive_ie
-                + profile_premis_ie
-                + profile_premis_rep
-            )
-
-        except Exception as e:
-            raise GraphParseError(f"Error when parsing graph: {str(e)}")
-
-        return data_graph
-
-    def _construct_shacl_graph(self) -> Graph:
-        shacl_graph = Graph()
-        shacl_graph.parse(str(self.shacl_sip()), format="turtle")
-        shacl_graph.parse(str(self.shacl_profile()), format="turtle")
-        return shacl_graph
