+++ app/resources/1.2/basic/sparql/descriptive_ie.sparql
@@ -23,7 +23,7 @@ CONSTRUCT
   }
 WHERE
   {
-    SERVICE <x-sparql-anything:location={{ bag_path }}/data/metadata/descriptive/dc_1.xml> {
+    SERVICE <x-sparql-anything:location={{ bag_path }}/data/metadata/descriptive/dc+schema.xml> {
         [] a fx:root;
             ?dc_1 [
                 a dct:identifier;
