+++ tests/resources/1.2/basic/sips/conform/data/mets.xml
@@ -21,7 +21,7 @@
 
     <!-- ref to descriptive metadata about IE -->
     <dmdSec ID="uuid-4d18fdda-a20c-4410-9fce-30c6bff14f50">
-        <mdRef LOCTYPE="URL" MDTYPE="DC" xlink:type="simple" xlink:href="metadata/descriptive/dc_1.xml" MIMETYPE="text/xml" SIZE="606" CREATED="2022-02-16T10:01:15.014+02:00" CHECKSUM="52e02444e8da664c660e2179a46847ed" CHECKSUMTYPE="MD5"/>
+        <mdRef LOCTYPE="URL" MDTYPE="DC" xlink:type="simple" xlink:href="metadata/descriptive/dc+schema.xml" MIMETYPE="text/xml" SIZE="606" CREATED="2022-02-16T10:01:15.014+02:00" CHECKSUM="52e02444e8da664c660e2179a46847ed" CHECKSUMTYPE="MD5"/>
     </dmdSec>
 
     <!-- ref to the PREMIS metadata about IE/subIE(s)/package -->
