+++ tests/resources/1.2/basic/sips/conform/manifest-md5.txt
@@ -1,5 +1,5 @@
-0609566a60e9da6f017f5593babbfee8  data/mets.xml
-52e02444e8da664c660e2179a46847ed  data/metadata/descriptive/dc_1.xml
+7974b51714e6131bd9b64cc1cb8f1a27  data/mets.xml
+52e02444e8da664c660e2179a46847ed  data/metadata/descriptive/dc+schema.xml
 c23db62c2abd690a500940fe42948b5b  data/metadata/preservation/premis.xml
 e03cc894b11a9d40ef7c3980474255c9  data/representations/representation_1/mets.xml
 02e4eb2f74ac06dd9cdfab861553697e  data/representations/representation_1/data/D523F963.jpg
