+++ tests/resources/1.2/basic/sips/conform_batch_id/bag-info.txt
@@ -1,3 +1,4 @@
 Bag-Software-Agent: bagit.py v1.8.1 <https://github.com/LibraryOfCongress/bagit-python>
-Bagging-Date: 2024-04-29
-Payload-Oxum: 14264.6
+Bagging-Date: 2024-04-30
+Payload-Oxum: 14269.6
+Meemoo-Batch-Identifier: batch-id
