+++ tests/resources/1.2/basic/sips/conform_batch_id/data/metadata/descriptive/dc+schema.xml
@@ -3,7 +3,7 @@
     xmlns:edtf="http://id.loc.gov/datatypes/edtf/"
     xmlns:xs="http://www.w3.org/2001/XMLSchema/"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-    xmlns="https://data.hetarchief.be/id/sip/1.0/basic">
+    xmlns="https://data.hetarchief.be/id/sip/1.2/basic">
 
     <dcterms:identifier>uuid-6b80a848-f2d2-44a4-bc55-56f272a8e5b4</dcterms:identifier>
     <dcterms:title xml:lang="nl">Titel</dcterms:title>
