+++ tests/resources/1.2/basic/sips/conform_extended/bag-info.txt
@@ -1,3 +1,3 @@
 Bag-Software-Agent: bagit.py v1.8.1 <https://github.com/LibraryOfCongress/bagit-python>
-Bagging-Date: 2024-04-29
-Payload-Oxum: 16482.6
+Bagging-Date: 2024-04-30
+Payload-Oxum: 16487.6
