+++ tests/resources/1.2/basic/sips/conform_extended/tagmanifest-md5.txt
@@ -1,3 +1,3 @@
-6e94862233cd693bce551b87bc95f96e bag-info.txt
-58134abeef4f81a4cd4344fda9a6588a manifest-md5.txt
+9edbfb795d60afceaede3edf90db2bf8 bag-info.txt
+387f8bc9f0dbc8c70c39dfe7cdf8fb79 manifest-md5.txt
 9e5ad981e0d29adc278f6a294b8c2aca bagit.txt
