+++ tests/resources/1.2/basic/sips/conform_local_ids/bag-info.txt
@@ -1,3 +1,3 @@
 Bag-Software-Agent: bagit.py v1.8.1 <https://github.com/LibraryOfCongress/bagit-python>
-Bagging-Date: 2024-04-29
-Payload-Oxum: 18678.6
+Bagging-Date: 2024-04-30
+Payload-Oxum: 18683.6
