+++ tests/resources/1.2/basic/sips/empty_graph/manifest-md5.txt
@@ -1,5 +1,5 @@
 ff29541763b75a32073b9bb757333220  data/mets.xml
-f71206007a0bad5fcd57280c1ce4181f  data/metadata/descriptive/dc_1.xml
+f71206007a0bad5fcd57280c1ce4181f  data/metadata/descriptive/dc+schema.xml
 3a43511c4797336e0fbde261c36e5dda  data/metadata/preservation/premis.xml
 b58f99f1a6c54016930fea0ff02645a6  data/representations/representation_1/mets.xml
 02e4eb2f74ac06dd9cdfab861553697e  data/representations/representation_1/data/D523F963.jpg
