+++ tests/resources/1.2/basic/sips/not_conform/manifest-md5.txt
@@ -1,5 +1,5 @@
 6267f24bec26e50a9d2aa1967eb81aeb  data/mets.xml
-69859860a63ab0aaf221a0a17a5b98e3  data/metadata/descriptive/dc_1.xml
+69859860a63ab0aaf221a0a17a5b98e3  data/metadata/descriptive/dc+schema.xml
 fda206a31ce2eb27c8713589a53bf0f7  data/metadata/preservation/premis.xml
 b439df2aa2bee594604570fef9d02b2d  data/representations/representation_1/mets.xml
 02e4eb2f74ac06dd9cdfab861553697e  data/representations/representation_1/data/D523F963.jpg
